# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VerifyMzansi API",
    "description": "Public and internal API for South Africa's verification-first marketplace.",
    "version": "1.0.0",
    "contact": {
      "name": "VerifyMzansi",
      "url": "https://verifymzansi.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "https://verifymzansi.com", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in with email/password",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/sign-out": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign out the current session",
        "operationId": "signOut",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Signed out successfully" }
        }
      }
    },
...
    "/api/admin/content/decide": {
      "post": {
        "tags": ["Admin"],
        "summary": "Moderate flagged content",
        "operationId": "decideContent",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contentId", "action"],
                "properties": {
                  "contentId": { "type": "string", "format": "uuid" },
                  "action": { "type": "string", "enum": ["approve", "remove", "warn"] },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Action taken" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie (set automatically on login)"
      }
    },
...
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  }
}